requirements.txt
Django==5.1.3
djangorestframework==3.15.2
influxdb-client==1.48.0
Pillow==9.0.1
urllib3~=1.26.5
requests~=2.28.2
cffi~=1.17.1
PySide6~=6.4.2
pip~=24.0
docutils~=0.19
sphinx~=6.1.3
keyring~=23.5.0
asgiref~=3.8.1
sqlparse~=0.4.3
colorama~=0.4.4
certifi~=2022.12.7
six~=1.16.0
python-dateutil~=2.8.1
reactivex~=4.0.4
setuptools~=62.6.0
Pygments~=2.14.0
pytz~=2022.7.1

python-dotenv~=1.1.0
flask~=3.1.0
paho-mqtt~=2.1.0
pygame~=2.1.2